Testrunning moved to model.Submission and made working again
This is step 1 towards resolving Issue #15.
The whole running tests and submitting testruns to the session is now
done in the submission model itself.
So the frontend controllers only have to get the output and show whats
relevant to the user.
Also this allows us to re-run testruns from the teachers point of view,
if needed.