New tests and fixes for model and url stuff
Tests and fixes for deliberately empty model object attributes.
Tests for *all* model instances that have an url attribute, which means
they map to a page which can be tested for correct rendering.
This of course may lead to duplicates with other test generators, but
better this than the other way round.

Fixes #197 partially.